+++ ansible/roles/ansible-role-conjur/Jenkinsfile
@@ -11,9 +11,9 @@ pipeline {
   stages {
     stage('Run tests') {
       steps {
-        sh './test.sh'
+        sh 'cd tests; ./test.sh'
 
-        // junit ''
+        junit 'tests/junit/*'
       }
     }
 
@@ -29,14 +29,7 @@ pipeline {
 
   post {
     always {
-      sh 'sudo chown -R jenkins:jenkins .'  // bad docker mount creates unreadable files TODO fix this
-      deleteDir()  // delete current workspace, for a clean build
-    }
-    failure {
-      slackSend(color: 'danger', message: "${env.JOB_NAME} #${env.BUILD_NUMBER} FAILURE (<${env.BUILD_URL}|Open>)")
-    }
-    unstable {
-      slackSend(color: 'warning', message: "${env.JOB_NAME} #${env.BUILD_NUMBER} UNSTABLE (<${env.BUILD_URL}|Open>)")
+      cleanupAndNotify(currentBuild.currentResult)
     }
   }
 }
