+++ ansible/roles/ansible-role-conjur/meta/.galaxy_install_info
@@ -1,1 +1,1 @@
-{install_date: 'Thu Sep 28 14:54:37 2017', version: master}
+{install_date: 'Sat Dec 16 08:33:58 2017', version: master}
