+++ ansible/roles/ansible-role-conjur/meta/main.yml
@@ -9,6 +9,7 @@ galaxy_info:
     - Dustin Collins (@dustinmm80)
     - Oren Ben Meir (@orenbm)
     - Jason Vanderhoof (@jvanderhoof)
+    - Kumbirai Tanekha (@doodlesbykumbi)
 
   min_ansible_version: '2.3'
