--- ansible/roles/ansible-role-conjur/molecule/.flake8
@@ -1,3 +1,0 @@
-[flake8]
-ignore = E128
-max-line-length = 120
