--- ansible/roles/ansible-role-conjur/molecule/configure-conjur-identity/playbook.yml
@@ -1,12 +1,0 @@
----
-- name: Converge
-  hosts: conjurized-container
-  become: true
-  roles:
-    - role: ansible-role-conjur/configure-conjur-identity
-      conjur_account: cucumber
-      conjur_appliance_url: "https://conjur-proxy-nginx"
-      conjur_host_factory_token: "{{lookup('env', 'HFTOKEN')}}"
-      conjur_ssl_certificate: "{{lookup('file', '../conjur.pem')}}"
-      conjur_host_name: "conjur_{{ ansible_hostname }}"
-      conjur_validate_certs: yes
