--- ansible/roles/ansible-role-conjur/molecule/retrieve-secret-in-master/playbook.yml
@@ -1,8 +1,0 @@
----
-- name: Converge
-  hosts: non-conjurized-container
-  tasks:
-    - name: Retrieve secret with master identity
-      vars:
-        super_secret_key: "{{lookup('retrieve_conjur_variable', 'ansible/master-password')}}"
-      shell: echo "{{super_secret_key}}" > ./conjur_secrets.txt
