--- ansible/roles/ansible-role-conjur/molecule/retrieve-secret-in-remote/molecule.yml
@@ -1,38 +1,0 @@
----
-dependency:
-  name: galaxy
-driver:
-  name: docker
-lint:
-  name: yamllint
-platforms:
-  - name: conjurized-container
-    image: conjurinc/cli5
-    dockercommands: true
-    entrypoint: sleep
-    command: infinity
-    networks:
-      - name: ansibleroleconjur_default
-    volumes:
-      - ../test-files/:/conjurinc/ansible/
-provisioner:
-  name: ansible
-  lint:
-    name: ansible-lint
-  playbooks:
-    create: ../create.yml
-    destroy: ../destroy.yml
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
-    options:
-      config-file: ../.flake8
-scenario:
-  name: retrieve-secret-in-remote
-  test_sequence:
-    - destroy
-    - create
-    - converge
-    - verify
-    - destroy
