--- ansible/roles/ansible-role-conjur/molecule/retrieve-secret-in-remote/playbook.yml
@@ -1,47 +1,0 @@
----
-- name: Converge
-  hosts: conjurized-container
-  become: true
-  vars:
-    CONJUR_ACCOUNT: "cucumber"
-    CONJUR_APPLIANCE_URL: "https://conjur-proxy-nginx"
-    CONJUR_CUSTOM_AUTHN_API_KEY: "{{lookup('env', 'CONJUR_CUSTOM_AUTHN_API_KEY')}}"
-    MOLECULE_TEMPLATES_DIRECTORY: "{{ lookup('env', 'MOLECULE_SCENARIO_DIRECTORY') }}/../templates"
-    CONJUR_SSL_CERTIFICATE: "{{lookup('file', '../conjur.pem')}}"
-  tasks:
-    - name: Place Conjur public SSL certificate
-      copy:
-        dest: /etc/conjur.pem
-        content: "{{ CONJUR_SSL_CERTIFICATE }}"
-        mode: 0644
-
-    - name: Symlink Conjur public SSL certificate into /etc/ssl/certs
-      file:
-        src: /etc/conjur.pem
-        dest: /etc/ssl/certs/conjur.crt
-        state: link
-      register: cert_symlink
-
-    - name: Rehash certs
-      command: 'c_rehash'
-      when: cert_symlink.changed
-
-    - name: Render /etc/conjur.conf
-      template:
-        src: "{{ MOLECULE_TEMPLATES_DIRECTORY}}/conjur.conf.j2"
-        dest: /etc/conjur.conf
-        mode: 0644
-
-    - name: Render /etc/conjur.identity
-      template:
-        src: "{{ MOLECULE_TEMPLATES_DIRECTORY}}/conjur.identity.j2"
-        dest: /etc/conjur.identity
-        mode: 0644
-
-    - name: Retrieve secrets with target identity
-      summon_conjur:
-        variables:
-          RETRIEVED_PASSWORD: "!var ansible/target-password"
-          ANOTHER_RETRIEVED_PASSWORD: "!var ansible/another-target-password"
-          LOCAL_VARIABLE: "local_variable_value"
-        command: /conjurinc/ansible/write_env_to_file.sh
