--- ansible/roles/ansible-role-conjur/molecule/templates/Dockerfile.non-conjurized-container
@@ -1,9 +1,0 @@
-
-RUN apt-get update \
-  && \
-  apt-get install -y \
-  python \
-  python-apt \
-  aptitude \
-  && \
-  rm -rf /var/lib/apt/lists/*
