--- ansible/roles/ansible-role-conjur/requirements-dev.txt
@@ -1,1 +1,0 @@
-molecule==2.0.4
