--- ansible/roles/ansible-role-conjur/test-files/fetch_cert.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-cd /conjurinc/ansible
-mkdir -p output
-rm -f output/conjur.pem
-rm -f /root/conjur-cucumber.pem
-rm -f /root/.conjurrc
-echo "yes" | conjur init -u "${CONJUR_PROXY_URL}" -a "${CONJUR_ACCOUNT}"
-cp /root/conjur-cucumber.pem output/conjur.pem
