--- ansible/roles/ansible-role-conjur/test-files/wait_for_server.sh
@@ -1,7 +1,0 @@
-#!/bin/bash -e
-
-for i in $(seq 20); do
-  curl -o /dev/null -fs -X OPTIONS ${CONJUR_APPLIANCE_URL} > /dev/null && echo "server is up" && break
-  echo "."
-  sleep 2
-done
