--- ansible/roles/ansible-role-conjur/test-files/write_env_to_file.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-env > ./conjur_env.txt
