+++ ansible/roles/ansible-role-conjur/tests/Dockerfile_nginx
