+++ ansible/roles/ansible-role-conjur/tests/ansible.cfg
@@ -1,0 +1,9 @@
+[defaults]
+host_key_checking = False
+error_on_undefined_vars = True
+timeout = 60
+inventory = inventory.tmp
+roles_path = ..
+remote_tmp = /tmp
+library = ../library
+lookup_plugins = ../lookup_plugins
