+++ ansible/roles/ansible-role-conjur/tests/inventory-playbook.yml
@@ -1,0 +1,6 @@
+---
+- name: Compile inventory template locally
+  hosts: localhost
+  tasks:
+    - name: compile inventory template
+      template: src=inventory.j2 dest=/conjurinc/ansible-role-conjur/tests/inventory.tmp
