+++ ansible/roles/ansible-role-conjur/tests/test_cases/retrieve-secret-in-master/playbook.yml
@@ -1,0 +1,8 @@
+---
+- name: Retrieve secret using master identity
+  hosts: testapp
+  tasks:
+    - name: Retrieve secret with master identity
+      vars:
+        super_secret_key: "{{lookup('retrieve_conjur_variable', 'ansible/master-password')}}"
+      shell: echo "{{super_secret_key}}" > ./conjur_secrets.txt
