+++ ansible/roles/ansible-role-conjur/tests/test_cases/retrieve-secret-in-remote/playbook.yml
@@ -1,0 +1,52 @@
+---
+- name: Retrieve secrets with remote target identity
+  hosts: testapp
+  become: true
+  vars:
+    CONJUR_ACCOUNT: "cucumber"
+    CONJUR_APPLIANCE_URL: "https://conjur-proxy-nginx"
+    CONJUR_CUSTOM_AUTHN_API_KEY: "{{lookup('env', 'CONJUR_CUSTOM_AUTHN_API_KEY')}}"
+    CONJUR_SSL_CERTIFICATE: "{{lookup('file', '../../conjur.pem')}}"
+  tasks:
+    - name: Place Conjur public SSL certificate
+      copy:
+        dest: /etc/conjur.pem
+        content: "{{ CONJUR_SSL_CERTIFICATE }}"
+        mode: 0644
+
+    - name: Symlink Conjur public SSL certificate into /etc/ssl/certs
+      file:
+        src: /etc/conjur.pem
+        dest: /etc/ssl/certs/conjur.crt
+        state: link
+      register: cert_symlink
+
+    - name: Rehash certs
+      command: 'c_rehash'
+      when: cert_symlink.changed
+
+    - name: Render /etc/conjur.conf
+      template:
+        src: "templates/conjur.conf.j2"
+        dest: /etc/conjur.conf
+        mode: 0644
+
+    - name: Render /etc/conjur.identity
+      template:
+        src: "templates/conjur.identity.j2"
+        dest: /etc/conjur.identity
+        mode: 0644
+
+    - name: Place write env to file script
+      copy:
+        dest: /write_env_to_file.sh
+        src: ./write_env_to_file.sh
+        mode: +x
+
+    - name: Retrieve secrets with target identity
+      summon_conjur:
+        variables:
+          RETRIEVED_PASSWORD: "!var ansible/target-password"
+          ANOTHER_RETRIEVED_PASSWORD: "!var ansible/another-target-password"
+          LOCAL_VARIABLE: "local_variable_value"
+        command: /write_env_to_file.sh
