+++ ansible/roles/ansible-role-conjur/tests/test_cases/retrieve-secret-in-remote/write_env_to_file.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+env > ./conjur_env.txt
